Add a notification for weather alerts
May need some tweaking as I do not know what valid states there are yet.
It might be "1" for there is an alert and "0" for there is not alert, or
there may be different numbers for different alerts. I haven't checked
code and may just wait until the present alert clears to see what it
changes to.